Translation of ## minggeJS Release v1.7
I was trying to translate it into English with the style as close as to the original document. There is, however, some Chinese wording I DO NOT really understand. Therefore, some words and sentences are translated directly and the grammar may be adjusted, in order to fit the original idea of the author, to the best of my understanding.